Classpath Improvements
Previously, the main application was not added to the classpath.  Instead only
the contents of Class-Path were.  This happened to work due to the current
working directory being the application root, but shouldn't have been
considered dependable.  Instead, this change ensures that the application root
is explicitly on the classpath.

Signed-off-by: Ben Hale <bhale@pivotal.io>