Do not use effect.NewExecutor use CommandExecutor instead
When running `syft`, do not use effect.NewExecutor which runs the command with a tty. This seems to cause an issue with syft (or possibly with our tty library pty), and in either case you end up with stray formatting characters even though we tell syft to be quiet. Using CommandExecutor is basically the same, but it doesn't run with a tty.

This also removes usage of ioutil.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>